refactor(api): promote new static implementation
